use-default-prompts hook
