+++ ui/client/src/components/agentic-ai/ExecutionTab.tsx
@@ -10,6 +10,7 @@ import { UmamiTrack } from '@/components/ui/umamitrack';
 import { UmamiEvents } from '@/config/umamiEvents';
 import { useSessionHub } from "@/hooks/use-session-hub";
 import { useCarouselLayout } from "@/hooks/use-carousel-layout";
+import { useDefaultPrompts } from "@/hooks/use-default-prompts";
 import { AnimatedPanelLayout } from "@/components/shared/AnimatedPanelLayout";
 import { AddFlowModal, DeleteSessionModal } from "@/components/shared/SessionModals";
 import { ViewModeToggle } from "@/components/shared/ViewModeToggle";
@@ -51,13 +52,7 @@ export default function ExecutionTab({ runId }: ExecutionTabProps): React.ReactE
 
   const isChatOnlyMode = hub.selectedSession?.fromSharedLink ?? false;
 
-  const defaultPrompts = useMemo(() => {
-    if (!hub.selectedSession?.blueprintId) return undefined;
-    const specDict = hub.blueprintSpecCache.get(hub.selectedSession.blueprintId);
-    const shortcuts = specDict?.prompt_shortcuts;
-    if (!Array.isArray(shortcuts)) return undefined;
-    return shortcuts.filter((p: PromptShortcut) => p.kind === "manual");
-  }, [hub.selectedSession?.blueprintId, hub.blueprintSpecCache]);
+  const defaultPrompts = useDefaultPrompts(hub);
 
   const carousel = useCarouselLayout({
     defaultChatPercent: 65,
